+++ CHANGELOG.md
@@ -32,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Improved VSCode Marketplace keywords.
+- Improved Visual Studio Code Marketplace keywords.
 
 ## [1.1.0] - 2023-08-29